+++ lib/core/channel/group/group_channel.g.dart
@@ -56,9 +56,9 @@ GroupChannel _$GroupChannelFromJson(Map<String, dynamic> json) {
     channelUrl: json['channel_url'] as String,
     name: json['name'] as String?,
     coverUrl: json['cover_url'] as String?,
-    creator: json['creator'] == null
+    creator: json['created_by'] == null
         ? null
-        : User.fromJson(json['creator'] as Map<String, dynamic>),
+        : User.fromJson(json['created_by'] as Map<String, dynamic>),
     createdAt: json['created_at'] as int?,
     data: json['data'] as String?,
     customType: json['custom_type'] as String?,
@@ -72,7 +72,7 @@ Map<String, dynamic> _$GroupChannelToJson(GroupChannel instance) =>
       'channel_url': instance.channelUrl,
       'name': instance.name,
       'cover_url': instance.coverUrl,
-      'creator': instance.creator?.toJson(),
+      'created_by': instance.creator?.toJson(),
       'created_at': instance.createdAt,
       'data': instance.data,
       'custom_type': instance.customType,